% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQTL_IEU.R
\name{PQTL_IEU}
\alias{PQTL_IEU}
\title{PQTL和在线结局MR分析}
\usage{
PQTL_IEU(savefile, PATH, GWASID, outname, kb, r2)
}
\arguments{
\item{savefile}{结果输出的文件夹名称}

\item{PATH}{切分好的文件夹}

\item{GWASID}{结局的GWASID号}

\item{outname}{结局的名称}

\item{kb}{聚类的距离}

\item{r2}{聚类的r2}
}
\description{
PQTL和在线结局MR分析
}
