% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQTL_local.R
\name{PQTL_local}
\alias{PQTL_local}
\title{PQTL和本地结局MR分析}
\usage{
PQTL_local(name, key, savefile, PATH, GWASsummay, outname, kb, r2)
}
\arguments{
\item{name}{学号}

\item{key}{密码}

\item{savefile}{结果输出的文件夹名称}

\item{PATH}{切分好的文件夹}

\item{GWASsummay}{结局的GWASsummary数据}

\item{outname}{结局的名称}

\item{kb}{聚类的距离}

\item{r2}{聚类的r2}
}
\description{
PQTL和本地结局MR分析
}
